티켓예매내역 수정

+++ MusicalTicket/app/src/main/java/com/gnuoynawh/musical/ticket/ui/TicketModifyActivity.kt
@@ -2,7 +2,6 @@ package com.gnuoynawh.musical.ticket.ui
 
 import android.app.Activity
 import android.content.Intent
-import android.graphics.BitmapFactory
 import android.os.Bundle
 import android.util.Log
 import android.view.View
@@ -14,6 +13,7 @@ import androidx.appcompat.widget.AppCompatEditText
 import androidx.appcompat.widget.AppCompatImageView
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.lifecycle.lifecycleScope
+import com.bumptech.glide.Glide
 import com.gnuoynawh.musical.ticket.R
 import com.gnuoynawh.musical.ticket.common.Constants
 import com.gnuoynawh.musical.ticket.db.MTicketDatabase
@@ -25,14 +25,14 @@ import java.nio.file.Files
 import java.text.SimpleDateFormat
 import java.util.*
 
-class AddTicketActivity: AppCompatActivity(), View.OnClickListener {
+class TicketModifyActivity: AppCompatActivity(), View.OnClickListener {
 
     private val db by lazy {
         MTicketDatabase.getDatabase(this)?.getTicket()
     }
 
-    private val btnAdd: AppCompatButton by lazy {
-        findViewById(R.id.btn_add)
+    private val btnModify: AppCompatButton by lazy {
+        findViewById(R.id.btn_modify)
     }
 
     private val btnTake: AppCompatButton by lazy {
@@ -51,24 +51,29 @@ class AddTicketActivity: AppCompatActivity(), View.OnClickListener {
         findViewById(R.id.iv_thumb)
     }
 
+    private val tvNumber: AppCompatTextView by lazy {
+        findViewById(R.id.tv_number)
+    }
+
     private val tvDate: AppCompatTextView by lazy {
         findViewById(R.id.tv_date)
     }
 
     private val edtTitle: AppCompatEditText by lazy { findViewById(R.id.edt_title) }
     private val edtSite: AppCompatEditText by lazy { findViewById(R.id.edt_site) }
-    private val edtNumber: AppCompatEditText by lazy { findViewById(R.id.edt_number) }
     private val edtCount: AppCompatEditText by lazy { findViewById(R.id.edt_count) }
     private val edtPlace: AppCompatEditText by lazy { findViewById(R.id.edt_place) }
 
     private var filePath = ""
     private var fileType = ""
 
+    private var ticket: Ticket? = null
+
     override fun onClick(v: View) {
 
         when(v.id) {
+            R.id.btn_modify -> updateData()
             R.id.btn_take -> takePicture()
-            R.id.btn_add -> insertData()
             R.id.btn_add_thumb -> addImage()
             R.id.btn_remove_thumb -> removeImage()
             R.id.tv_date -> {
@@ -84,65 +89,53 @@ class AddTicketActivity: AppCompatActivity(), View.OnClickListener {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_add_ticket)
+        setContentView(R.layout.activity_modify_ticket)
 
-        btnAdd.setOnClickListener(this)
+        btnModify.setOnClickListener(this)
         btnTake.setOnClickListener(this)
         btnAddThumb.setOnClickListener(this)
         btnRemoveThumb.setOnClickListener(this)
         tvDate.setOnClickListener(this)
+
+        ticket = intent.getSerializableExtra("ticket") as Ticket
+        setData()
+
     }
 
-    private fun verify() : Boolean {
-        when {
-            edtTitle.text.isNullOrEmpty() -> {
-                Toast.makeText(this, "제목을 입력하세요.", Toast.LENGTH_SHORT).show()
-                return true
-            }
-            edtSite.text.isNullOrEmpty() -> {
-                Toast.makeText(this, "예매처를 입력하세요.", Toast.LENGTH_SHORT).show()
-                return true
-            }
-            edtNumber.text.isNullOrEmpty() -> {
-                Toast.makeText(this, "예매번호를 입력하세요.", Toast.LENGTH_SHORT).show()
-                return true
-            }
-            tvDate.text.isNullOrEmpty() -> {
-                Toast.makeText(this, "관람일을 입력하세요.", Toast.LENGTH_SHORT).show()
-                return true
-            }
-            edtPlace.text.isNullOrEmpty() -> {
-                Toast.makeText(this, "장소를 입력하세요.", Toast.LENGTH_SHORT).show()
-                return true
-            }
-            edtCount.text.isNullOrEmpty() -> {
-                Toast.makeText(this, "매수를 입력하세요.", Toast.LENGTH_SHORT).show()
-                return true
-            }
-            else -> {}
-        }
+    private fun setData() {
+        edtTitle.setText(ticket?.title)
+        edtSite.setText(ticket?.site)
+        tvNumber.text = ticket?.number
+        tvDate.text = ticket?.date
+        edtPlace.setText(ticket?.place)
+        edtCount.setText(ticket?.count)
 
-        return false
+        filePath = ticket?.image.toString()
+        fileType = ticket?.imageType.toString()
+
+        Glide.with(this)
+            .load(
+                if (ticket?.imageType == Constants.IMAGE_TYPE_FILE)
+                    File(filePath)
+                else
+                    filePath
+            ).error(R.color.black)
+            .into(ivThumb)
     }
 
-    private fun insertData() {
+    private fun updateData() {
+        ticket?.title = edtTitle.text.toString()
+        ticket?.site = edtSite.text.toString()
+        ticket?.number = tvNumber.text.toString()
+        ticket?.date = tvDate.text.toString()
+        ticket?.place = edtPlace.text.toString()
+        ticket?.count = edtCount.text.toString()
 
-        if (verify())
-            return
-
-        val ticket = Ticket()
-        ticket.title = edtTitle.text.toString()
-        ticket.site = edtSite.text.toString()
-        ticket.number = edtNumber.text.toString()
-        ticket.date = tvDate.text.toString()
-        ticket.place = edtPlace.text.toString()
-        ticket.count = edtCount.text.toString()
-
-        ticket.image = filePath
-        ticket.imageType = fileType
+        ticket?.image = filePath
+        ticket?.imageType = fileType
 
         lifecycleScope.launch {
-            db?.insert(ticket)
+            db?.insert(ticket!!)
             finish()
         }
     }
@@ -164,8 +157,10 @@ class AddTicketActivity: AppCompatActivity(), View.OnClickListener {
     private fun removeImage() {
         ivThumb.setImageResource(0)
 
-        val file = File(filePath)
-        file.delete()
+        if (ticket?.imageType == Constants.IMAGE_TYPE_FILE) {
+            val file = File(filePath)
+            file.delete()
+        }
 
         filePath = ""
         fileType = ""
@@ -175,10 +170,17 @@ class AddTicketActivity: AppCompatActivity(), View.OnClickListener {
         if (it.resultCode == Activity.RESULT_OK) {
 
             val path = it.data?.getStringExtra("filePath")
-            ivThumb.setImageBitmap(BitmapFactory.decodeFile(path))
+            path?.let {
 
-            filePath = path ?: ""
-            fileType = Constants.IMAGE_TYPE_FILE
+                Glide.with(this)
+                    .load(File(path))
+                    .error(R.color.black)
+                    .into(ivThumb)
+
+                filePath = path
+                fileType = Constants.IMAGE_TYPE_FILE
+
+            }
         }
     }
 
@@ -191,7 +193,10 @@ class AddTicketActivity: AppCompatActivity(), View.OnClickListener {
             Files.copy(file.toPath(), outputFile.toPath());
 
             val path = outputFile.path
-            ivThumb.setImageBitmap(BitmapFactory.decodeFile(path))
+            Glide.with(this)
+                .load(File(path))
+                .error(R.color.black)
+                .into(ivThumb)
 
             filePath = path ?: ""
             fileType = Constants.IMAGE_TYPE_FILE
